Added Seller-Auth

+++ EsyaStoreApi/Controllers/ProductsController.cs
@@ -1,6 +1,8 @@
-﻿using EsyaStore.Data.Context;
+﻿using System.Security.Claims;
+using EsyaStore.Data.Context;
 using EsyaStore.Data.Entity;
 using EsyaStoreApi.Model;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -40,9 +42,28 @@ namespace EsyaStoreApi.Controllers
 
         }
 
+        [HttpGet("seller/{Id}")]
+        public async Task<IActionResult> GetProductsBySeller(int Id)
+        {
+            var sellerClaim = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (sellerClaim != Id.ToString())
+            {
+                return Unauthorized("You are not authorized to add product for this seller.");
+            }
+            var product=await _context.products.Where(p=>p.SellerId==Id).ToListAsync();
+            if (product is null) { return NotFound(); };
+            return Ok(product);
+        }
+
         [HttpPost]
+        [Authorize(Policy ="SellerPolicy")]
         public async Task<IActionResult> PostProduct(CreateProductDTO newProduct)
         {
+            var sellerClaim = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if(sellerClaim != newProduct.SellerId.ToString())
+            {
+                return Unauthorized("You are not authorized to add product for this seller.");
+            }
             var product = new Products
             {
                 ProductName = newProduct.ProductName,
@@ -62,9 +83,15 @@ namespace EsyaStoreApi.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> EditProduct(int id,UpdateProductDTO editProduct) { 
+        [Authorize(Policy ="SellerPolicy")] 
+        public async Task<IActionResult> EditProduct(int id,UpdateProductDTO editProduct) {
+            var sellerClaim = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
             var product=await _context.products.FindAsync(id);
             if(product is null) { return NotFound(); }
+            if (sellerClaim != product.SellerId.ToString())
+            {
+                return Unauthorized("You are not authorized to edit this product.");
+            }
             product.ProductName=editProduct.ProductName;
             product.ProductDescription=editProduct.ProductDescription;
             product.ProductCategory=editProduct.ProductCategory;
@@ -80,9 +107,15 @@ namespace EsyaStoreApi.Controllers
         }
 
         [HttpDelete("{id}")]
+        [Authorize(Policy ="SellerPolicy")]
         public async Task<IActionResult> DeleteProduct(int id) {
+            var sellerClaim = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
             var product=await _context.products.FindAsync(id);
             if (product is null) { return NotFound(); }
+            if (sellerClaim != product.SellerId.ToString())
+            {
+                return Unauthorized("You are not authorized to edit this product.");
+            };
             _context.products.Remove(product);
             await _context.SaveChangesAsync();
             return NoContent();
